refactor: remove basedto + page-reponse
