--- client/src/types/base-dto.ts
@@ -1,3 +1,0 @@
-export interface BaseDto {
-  id: number;
-}
