--- client/src/types/page-response.ts
@@ -1,6 +1,0 @@
-export interface PageResult<T> {
-  count: number;
-  next: string;
-  previous: string;
-  results: T[];
-}
